requirements.txt
PyYaml
requests
jsonpath
pyyaml-include==1.3.2
flask_cors
pymysql
flask
flask_jwt_extended
werkzeug
flask_script
